extract report data to hook
